Refactor: Rename Redux to Reducer to avoid confusion
